requirements.txt
# Main Backend Dependencies
fastapi>=0.115
uvicorn>=0.34
pydantic[email]>=2.10
python-dotenv>=1.0.1
sqlalchemy>=2.0.39
psycopg2-binary>=2.9.10
alembic>=1.15.1
python-multipart>=0.0.20
pydantic-settings>=2.8.1
boto3>=1.37.33
sqladmin[full]>=0.20.1
itsdangerous>=2.2.0
gunicorn>=23.0.0
openai>=1.82.0
posthog>=4.2.0
pillow>=11.2.1
resend>=2.10.0
celery>=5.5.3
stripe>=12.3.0
redis>=5.0.1

# PDF Processing Dependencies
pymupdf>=1.25.5
pymupdf4llm>=0.0.25
markitdown[pdf]>=0.1.2

# Development Dependencies
types-psycopg2>=2.9.21
types-requests>=2.32.0
sqlalchemy2-stubs>=0.0.2a38
types-boto3>=1.37.33
celery-stubs>=0.1.3
types-psutil>=7.0.0

# Optional Dependencies
google-genai>=1.7.0  # For Google AI integration
firecrawl-py>=2.7.0  # For web crawling 